Moved get alt_profiles call to utils
